[Core]support UI
